requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
plexapi>=4.15.4
python-dateutil>=2.8.2

# Optional dependencies for development and testing
pytest>=7.4.0
pylint>=2.17.5
black>=23.7.0

# Monitoring and logging
prometheus_client>=0.17.1
python-json-logger>=2.0.7